Adjust examples
